feat(nushell): fix *nix

+++ windows/AppData/Roaming/nushell/config.nu
@@ -17,13 +17,15 @@ if (which mise | is-not-empty) {
     rm -f $mise_config
 }
 
-let starship_config = $nu.data-dir | path join vendor autoload starship.nu
-if (which starship | is-not-empty) {
-    if not ($starship_config | path exists) {
-        starship init nu | save -f $starship_config
+if ($nu.os-info | get name) == "windows" {
+    let starship_config = $nu.data-dir | path join vendor autoload starship.nu
+    if (which starship | is-not-empty) {
+        if not ($starship_config | path exists) {
+            starship init nu | save -f $starship_config
+        }
+    } else {
+        rm -f $starship_config
     }
-} else {
-    rm -f $starship_config
 }
 
 let zoxide_config = $nu.data-dir | path join vendor autoload zoxide.nu
